Add more styling, fix form validation and add intro to Quiz
Added form validation to first and last name of the signup form.

Added introductory note about the questions in the Quiz for user.

Took the PR from Josh and merged it into this PR. In order to get
the latest code up to the repo, I needed to merge on my own. Jason
submitted a PR that had fixed all file structure issues. Josh then
submitted a PR while Jason's PR was still pending. I merged Jason's
and then Josh's PR was unable to be automatically merged. I closed
that PR and did the merging myself manually.

Resolves: #70, $74
Fixes: #69