Added input for email of partner on signup (#81)
* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Refactor routing on singup and after completing quiz

* Add browserHistory feature to formSubmit on Quiz component

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Refactor routing on singup and after completing quiz

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Removed typos from Quiz component

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Begin singup form validation and refactor

* Add more styling, fix form validation and add intro to Quiz

Added form validation to first and last name of the signup form.

Added introductory note about the questions in the Quiz for user.

Took the PR from Josh and merged it into this PR. In order to get
the latest code up to the repo, I needed to merge on my own. Jason
submitted a PR that had fixed all file structure issues. Josh then
submitted a PR while Jason's PR was still pending. I merged Jason's
and then Josh's PR was unable to be automatically merged. I closed
that PR and did the merging myself manually.

Resolves: #70, $74
Fixes: #69

* Add input field for partner email on signup

Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68

* Add input field for partner email on signup
Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68