Calendar Feature MVP (#93)
* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Refactor routing on singup and after completing quiz

* Add browserHistory feature to formSubmit on Quiz component

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Refactor routing on singup and after completing quiz

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Removed typos from Quiz component

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Begin singup form validation and refactor

* Add more styling, fix form validation and add intro to Quiz

Added form validation to first and last name of the signup form.

Added introductory note about the questions in the Quiz for user.

Took the PR from Josh and merged it into this PR. In order to get
the latest code up to the repo, I needed to merge on my own. Jason
submitted a PR that had fixed all file structure issues. Josh then
submitted a PR while Jason's PR was still pending. I merged Jason's
and then Josh's PR was unable to be automatically merged. I closed
that PR and did the merging myself manually.

Resolves: #70, $74
Fixes: #69

* Add input field for partner email on signup

Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68

* Add input field for partner email on signup
Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68

* Fixed style guid errors on Signup.js

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Continue refactor of back end and front end for MVP

* Add logic on back-end for updating score

Need to work on using dynamic data, but the connection is there

* Continue refactor of back end and front end for MVP

* Begin singup form validation and refactor

* Add more styling, fix form validation and add intro to Quiz

Added form validation to first and last name of the signup form.

Added introductory note about the questions in the Quiz for user.

Took the PR from Josh and merged it into this PR. In order to get
the latest code up to the repo, I needed to merge on my own. Jason
submitted a PR that had fixed all file structure issues. Josh then
submitted a PR while Jason's PR was still pending. I merged Jason's
and then Josh's PR was unable to be automatically merged. I closed
that PR and did the merging myself manually.

Resolves: #70, $74
Fixes: #69

* Add input field for partner email on signup

Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68

* Add input field for partner email on signup
Also refactored and compressed client side code to clean up
some style guide errors.

Resolves: #68

* Begin work on calendar and create new event

* Begin work on calendar feature -- front-end

Create new link to display calendar
Add button to create new events (still work in progress)
Begin redux implementation

* Begin work on parsing date from events and date inputs on form

* Fix bug on calendar rendering with time zones

* Complete first version of calendar feature. Need to add database

* Clean up syntax and code using ES6 string templating

* Fix typo in quizActions that prevented Quiz from loading

* Begin work on integrating database

Bug with the SQL query in the events API route

* Finish calendar feature and connected to database

Works on signup, but bug on singin with coupleID being undefined

* Remove console.log statements

* Post rebase refactor to fix some errors with Authentication

* Complete database refactor and code cleanup