Add more test support, including Coveralls (#114)
* Prepare for meter tests

Added directory and files for future tests on the meter component.

* Add quick fix for validate colors

Added a quick CSS fix for neutralizing the validation input colors that
come from Materialize. But this isn't a complete solution. A lot of
overrides are happening between all the components as well as
Materialize.

* Add Coveralls

Added Coveralls integration with Travis-CI.

* Add build status and coverage buttons

Added code in README for showing Travis-CI and Coveralls buttons.

* Fix Travi-CI

An attempt to fix failed build.